Update patch set 2
Patch Set 2:

This change is on the borderline of where automated testing is practical.

Use of SWTbot or Jubula or ... remains a to-do for the project as a whole, so UI testing is limited to a few tests that confiorm that the Xtext editors/console offer some plausible user experience. This is very useful given the decreasing stability of platform APIs.

It should be possible to contruct a JUnit test that puts up a new file page validates the initial page content, clicks next and validates the subsequent file creation and editor opening. If really necessary, a couple of for-test-only read-only global variables might assist the JUnit test in detecting what has happened.

Patch-set: 2