Update patch set 1
Patch Set 1:

It is not a good idea to drop an enhancement on to a project without prior discussion, since the project invariably has its way of doing things and often has greater understanding about the way the project works and can be extended while improving rather than degrading its integrity.

Once we have an agreed design approach we can start to consider implementation.

Patch-set: 1