Update patch set 1
Patch Set 1:

The change looks sensible, but in order to maintain 100% compatibility the design policy changed.

Rather than a constraint evaluation optionally returning a TupleValue (not 100% compatible), it may instead return the status part projection of a TupleValue (100% compatible). Smart tooling may recognise the status part projection idiom to discover additional sibling results. Differently smart tooling may recognise that some of the TupleValue parts are only used when the status is false and not compute them when the status is true.

There is probably at least a misleading comment to fix.

How did you get hold of a TupleValue return?

Patch-set: 1